refactor: make modifier extensions in one file
